pyproject.toml
[tool.poetry]
name = "weather_app_cookies"
version = "0.1.0"
description = ""
authors = ["jeas560 <jeas560@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
fastapi = {extras = ["all"], version = "^0.108.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
responses = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
